### 2024-05-15 #6743 @sandeshkatakam
> @dirmeier @yasinkutuk @sandeshkatakam
>
> would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
>

Sure, I will be able to do it
