+++ .github/workflows/CI.yml
@@ -58,7 +58,7 @@ jobs:
       - uses: julia-actions/julia-processcoverage@v1
         if: contains(fromJson('["1", "1.9"]'), matrix.version) && matrix.os == 'ubuntu-latest'
         continue-on-error: true
-      - uses: codecov/codecov-action@v4.3.1
+      - uses: codecov/codecov-action@v4.4.0
         if: contains(fromJson('["1", "1.9"]'), matrix.version) && matrix.os == 'ubuntu-latest'
         with:
           file: lcov.info
