### 2024-05-21 #6743 @fabian-s
hi @dirmeier @yasinkutuk @odunbar

would any of you be willing to review this submission for JOSS?
