### 2024-05-27 #6743 @fabian-s
@avik-pal @CeciliaCoelho
would any of you be willing to review this submission for JOSS?
