### 2024-05-27 #6743 @avik-pal
> @avik-pal @CeciliaCoelho would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html)

Yes I can review this submission
