### 2024-05-28 #6794 @editorialbot
[REVIEW]: AutoEncoderToolkit.jl: A Julia package for training (Variational) Autoencoders
**Submitting author:** @mrazomej (<a href="http://orcid.org/0000-0002-9510-0527">Manuel Razo-Mejia</a>)
**Repository:** https://github.com/mrazomej/AutoEncoderToolkit.jl
**Branch with paper.md** (empty if default branch): publication
**Version:** v0.1.2
