### 2024-05-28 #6794 @fabian-s
@mrazomej:

while the reviewers start working through their checklists, please add DOIs for your references, where possible:

![image](https://github.com/openjournals/joss-reviews/assets/998541/d6339472-0c71-43ad-8c4c-e2d61204ad4a)
