### 2024-05-28 917bb24
add missing DOIs to references
