+++ paper.bib
@@ -6,8 +6,7 @@
   eprint = {1710.11379},
   eprinttype = {arxiv},
   eprintclass = {stat},
-  url = {http://arxiv.org/abs/1710.11379},
-  urldate = {2023-03-28},
+  doi = {10.48550/arXiv.1710.11379},
   pubstate = {preprint}
 }
 
@@ -15,7 +14,8 @@
   title = {Hamiltonian {{Variational Auto-Encoder}}},
   author = {Caterini, Anthony L and Doucet, Arnaud and Sejdinovic, Dino},
   date = {2018},
-  pages = {11}
+  pages = {11},
+  doi = {10.48550/arXiv.1805.11328}
 }
 
 @online{chadebec2020,
@@ -25,8 +25,7 @@
   eprint = {2010.11518},
   eprinttype = {arxiv},
   eprintclass = {cs, math, stat},
-  url = {http://arxiv.org/abs/2010.11518},
-  urldate = {2023-11-09},
+  doi = {10.48550/arXiv.2010.11518},
   pubstate = {preprint}
 }
 
@@ -37,8 +36,7 @@
   eprint = {2209.07370},
   eprinttype = {arxiv},
   eprintclass = {cs, stat},
-  url = {http://arxiv.org/abs/2209.07370},
-  urldate = {2023-11-09},
+  doi = {10.48550/arXiv.2209.07370},
   pubstate = {preprint}
 }
 
@@ -95,8 +93,7 @@
   eprint = {1312.6114},
   eprinttype = {arxiv},
   eprintclass = {cs, stat},
-  url = {http://arxiv.org/abs/1312.6114},
-  urldate = {2022-11-21},
+  doi = {10.48550/arXiv.1312.6114},
   pubstate = {preprint}
 }
 
@@ -111,10 +108,16 @@
 
 @article{lopez2018,
   title = {Deep Generative Modeling for Single-Cell Transcriptomics},
-  author = {Lopez, Romain},
-  date = {2018},
+  author = {Lopez, Romain and Regier, Jeffrey and Cole, Michael B. and Jordan, Michael I. and Yosef, Nir},
+  date = {2018-12},
   journaltitle = {Nature Methods},
-  volume = {15}
+  shortjournal = {Nat Methods},
+  volume = {15},
+  number = {12},
+  pages = {1053--1058},
+  publisher = {Nature Publishing Group},
+  issn = {1548-7105},
+  doi = {10.1038/s41592-018-0229-2}
 }
 
 @online{rezaabad2020,
@@ -124,8 +127,7 @@
   eprint = {1912.13361},
   eprinttype = {arxiv},
   eprintclass = {cs, stat},
-  url = {http://arxiv.org/abs/1912.13361},
-  urldate = {2023-11-07},
+  doi = {10.48550/arXiv.1912.13361},
   pubstate = {preprint}
 }
 
@@ -137,7 +139,6 @@
   eprint = {1706.02262},
   eprinttype = {arxiv},
   eprintclass = {cs, stat},
-  url = {http://arxiv.org/abs/1706.02262},
-  urldate = {2022-12-08},
+  doi = {10.48550/arXiv.1706.02262},
   pubstate = {preprint}
 }
