### 2024-05-28 #6794 @mrazomej
> @mrazomej:
>
> while the reviewers start working through their checklists, please add DOIs for your references, where possible:

@fabian-s ,
I added a `DOI` for all papers with one available. However, I couldn't find a `DOI` anywhere for the following papers:
> 1. Chen, N. et al. Metrics for Deep Generative Models. in Proceedings of the Twenty-First International Conference on Artificial Intelligence and Statistics 1540–1550 (PMLR, 2018).
> 2. Higgins, I. et al. β-VAE: Learning Basic Visual Concepts with a Constrained Variational Framework. (2017).
