### 2024-06-24 #6794 @avik-pal
> @avik-pal we notice you have not edited your reviewer checklist in a while. please let us know your timeframe for completing the review of this submission. JOSS aims for a collaborative style of review, so if you come across any open problems, please let the authors know in this thread or by opening up issues in their repo.

I plan to finish this within the next 2 weeks.
