### 2024-07-07 mrazomej/AutoEncoderToolkit.jl#14 @avik-pal
General Feedback from JOSS reviews
1. Start with atleast a `0.1.*` versioning. `0.0.*` means every change is breaking
2. Caret specifiers are not needed
3. CUDA Support:
	1. Add CUDA tests. buildkite might be hard for repos in personal accounts, but since the cuda code in the ext involves custom kernels, I would expect tests at the bare minimum.
4. https://github.com/mrazomej/AutoEncoderToolkit.jl/blob/49653858a1e0bb70eb08bd27a18875c59a7c4491/src/layers.jl#L192. Flux now prefers `@layer`
5. Coverage for the project seems extremely low, even excluding the cuda tests it is roughly 60%.
6. > Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

  Can a section for this be added?
7. Pattern like https://github.com/mrazomej/AutoEncoderToolkit.jl/blob/49653858a1e0bb70eb08bd27a18875c59a7c4491/src/decoders.jl#L452-L456 makes your entire code type-unstable, which might be okay for large networks, but best to avoid that pattern.
