### 2024-07-07 #6794 @avik-pal
Overall paper and software looks good. Remaining issues are highlighted in https://github.com/mrazomej/AutoEncoderToolkit.jl/issues/14
