### 2024-07-07 6d1c723
remove experimental diffgeo files
