+++ src/diffgeo/diffeq.jl
@@ -1,331 +1,0 @@
-# ==============================================================================
-# Geodesic Equation computation via differential equations
-# ==============================================================================
-
-# ------------------------------------------------------------------------------
-# Derivative of the Riemannian metric with respect to the coordinates in the
-# input space
-# ------------------------------------------------------------------------------
-
-@doc raw"""
-    ∂M̲̲∂γ̲(manifold, val, out_dim)
-
-Compute the derivative of the Riemannian metric `M̲̲` with respect to the
-coordinates in the input space using `Zygote.jl`.
-
-# Arguments
-- `manifold::Function`: A function defining the manifold.
-- `z::AbstractVector`: A vector specifying the point on the manifold where the
-  derivative should be evaluated.
-- `out_dim::Int`: Dimensionality of the output space. This is essential because
-  `Zygote.jl` can't compute certain derivatives automatically. We must iterate
-  over each dimension of the output space manually.
-
-# Returns
-- `∂M̲̲::Array{<:Number}`: A rank-3 tensor evaluating the derivative of
-  the Riemannian manifold metric.
-
-# Example
-```julia-repl
-manifold(x) = [x[1]^2 + x[2], x[1]*x[2]]
-val = [1.0, 2.0]
-∂M = ∂M̲̲∂γ̲(manifold, val, 2)
-```
-"""
-function ∂M̲̲∂γ̲(
-    manifold::Function, z::AbstractVector, out_dim::Int
-)
-    # Compute the manifold Jacobian. This should be a D×d matrix where D is the
-    # dimension of the output space, and d is the dimension of the manifold.
-    # Note that we use first() to extract the object we care about from the
-    # Zygote output.
-    J̲̲ = first(Zygote.jacobian(manifold, z))
-
-    # Compute Hessian tensor, i.e., the tensor with the manifold second
-    # derivatives. This should be a D×d×d third-order tensor, where D is the
-    # dimension of the output space, and d is the dimension of the manifold.
-    # Note that we have to manually evaluate the hessian on each dimension of
-    # the output.
-    H̲̲ = permutedims(
-        cat(
-            [Zygote.hessian(v -> manifold(v)[D], z) for D = 1:out_dim]...,
-            dims=3
-        ),
-        (3, 1, 2)
-    )
-
-    # Compute the derivative of the Riemmanian metric. This should be a d×d×d
-    # third-order tensor.
-    return @tensor ∂M̲̲[i, j, k] := H̲̲[l, i, k] * J̲̲[l, j] +
-                                    H̲̲[l, k, j] * J̲̲[l, i]
-end # function
-
-# ------------------------------------------------------------------------------
-
-@doc raw"""
-    ∂M̲̲∂γ̲(manifold, z, out_dim)
-
-Compute the derivative of the Riemannian metric `M̲̲` with respect to the
-coordinates in the input space using `Zygote.jl`.
-
-# Arguments
-- `manifold::Flux.Chain`: A neural network model (as a chain) defining the
-  manifold.
-- `z::AbstractVector`: A vector specifying the input to the manifold (neural
-  network) where the derivative should be evaluated.
-- `out_dim::Int`: Dimensionality of the output space. This is essential because
-  `Zygote.jl` can't compute certain derivatives automatically. We must iterate
-  over each dimension of the output space manually.
-
-# Returns
-- `∂M̲̲::Array{<:Number}`: A rank-3 tensor evaluating the derivative of
-  the Riemannian manifold metric.
-
-# Example
-```julia-repl
-model = Chain(Dense(2, 3, relu), Dense(3, 2))
-val = [1.0, 2.0]
-∂M = ∂M̲̲∂γ̲(model, val, 2)
-```
-"""
-function ∂M̲̲∂γ̲(
[…]
