### 2024-07-07 6717e0b
add community guidelines section
