+++ docs/src/guidelines.md
@@ -1,0 +1,22 @@
+# Community Guidelines
+
+## Contributing to the Software
+For those interested in contributing to AutoEncoderToolkit.jl, please refer to
+the [GitHub repository](https://github.com/mrazomej/AutoEncoderToolkit.jl). The
+project welcomes contributions to 
+
+- Expand the list of available models.
+- Improve the performance of existing models.
+- Add new features to the toolkit.
+- Improve the documentation.
+
+## Reporting Issues or Problems
+If you encounter any issues or problems with the software, you can report them
+directly on the [GitHub repository's issues
+page](https://github.com/mrazomej/AutoEncoderToolkit.jl/issues).
+
+## Seeking Support
+For support and further inquiries, consider checking the documentation and
+existing issues on the GitHub repository. If you still do not find the answer,
+you can open a new issue on the [GitHub repository's issues
+page](https://github.com/mrazomej/AutoEncoderToolkit.jl/issues).
