### 2024-07-07 mrazomej/AutoEncode.jl#14 @mrazomej
@avik-pal, thank you so much for taking the time to carefully review the paper and the software. I agree with all comments and suggestions and will promptly implement the necessary changes.

I would like to follow up with some questions and clarifications:

> 2. Caret specifiers are not needed

How would you rewrite this example from the `Project.toml` file without the caret specifiers?
```
[compat]
CUDA = "^5.3.3"
ChainRulesCore = "^1.23.0"
```

> 3. CUDA Support:
> Add CUDA tests. buildkite might be hard for repos in personal accounts, but since the cuda code in the ext involves custom kernels, I would expect tests at the bare minimum.

Would it suffice to have a test file that runs locally on my `CUDA`-available machine?

> 7. Pattern like
>
> [AutoEncoderToolkit.jl/src/decoders.jl](https://github.com/mrazomej/AutoEncoderToolkit.jl/blob/49653858a1e0bb70eb08bd27a18875c59a7c4491/src/decoders.jl#L452-L456)
> Lines 452 to 456 in [4965385](https://github.com/mrazomej/AutoEncoderToolkit.jl/commit/49653858a1e0bb70eb08bd27a18875c59a7c4491)
```julia
 struct JointGaussianLogDecoder <: AbstractGaussianLogDecoder
     decoder::Flux.Chain
     µ::Flux.Dense
     logσ::Flux.Dense
 end
```
> makes your entire code type-unstable, which might be okay for large networks, but best to avoid that pattern.

Would rewriting the `struct` as
```julia
struct JointGaussianLogDecoder{D<:Flux.Chain,L<:Flux.Dense} <: AbstractGaussianLogDecoder
    decoder::D
    µ::L
    logσ::L
end
```
fix the type instability?

Thanks again for your help!
