### 2024-07-07 mrazomej/AutoEncode.jl#14 @avik-pal
```julia
[compat]
CUDA = "5.3.3"
ChainRulesCore = "1.23.0"
```

is same as specifying `^...`.

> Would it suffice to have a test file that runs locally on my CUDA-available machine?

Yes, you can commit that. Add a explicit check for `CUDA.functional()` and run those test files if  cuda is functional (which would be false for github actions)

> fix the type instability?

Yes. You can also use https://github.com/jonniedie/ConcreteStructs.jl which makes it easier to write these things.
