+++ Project.toml
@@ -26,20 +26,20 @@ cuDNN = "02a925ec-e4fe-4b08-9a7e-0d78e3d38ccd"
 AutoEncoderToolkitCUDAExt = "CUDA"
 
 [compat]
-CUDA = "^5.3.3"
-ChainRulesCore = "^1.23.0"
-Clustering = "^0.15.7"
-Distances = "^0.10.11"
-Distributions = "^0.25.108"
-EllipsisNotation = "^1.8.0"
-Flux = "^0.14.15"
-ForwardDiff = "^0.10.36"
-NearestNeighbors = "^0.4.16"
-StatsBase = "^0.34.3"
-TaylorDiff = "^0.2.1"
-Zygote = "^0.6.69"
-cuDNN = "^1.3.1"
-julia = "^1.9"
+CUDA = "5.3.3"
+ChainRulesCore = "1.23.0"
+Clustering = "0.15.7"
+Distances = "0.10.11"
+Distributions = "0.25.108"
+EllipsisNotation = "1.8.0"
+Flux = "0.14.15"
+ForwardDiff = "0.10.36"
+NearestNeighbors = "0.4.16"
+StatsBase = "0.34.3"
+TaylorDiff = "0.2.1"
+Zygote = "0.6.69"
+cuDNN = "1.3.1"
+julia = "1.9"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
