### 2024-07-07 bd98730
add ConcreteStructs and make encoders and decoders type stable
