+++ src/decoders.jl
@@ -3,6 +3,8 @@ import Flux
 import TaylorDiff
 # Import library to use Ellipsis Notation
 using EllipsisNotation
+# Import ConcreteStructs module
+using ConcreteStructs: @concrete
 
 ## ============================================================================
 # Abstract Decoder Types
@@ -122,7 +124,7 @@ abstract type AbstractGaussianLinearDecoder <: AbstractGaussianDecoder end
 # ==============================================================================
 
 @doc raw"""
-`struct Decoder`
+`struct Decoder{D<:Flux.Chain} <: AbstractDeterministicDecoder`
 
 Default decoder function for deterministic autoencoders. The `decoder` network
 is used to map the latent space representation directly back to the original
@@ -137,12 +139,21 @@ data space.
 dec = Decoder(Flux.Chain(Dense(20, 400, relu), Dense(400, 784)))
 ```
 """
-struct Decoder <: AbstractDeterministicDecoder
-    decoder::Flux.Chain
+@concrete struct Decoder <: AbstractDeterministicDecoder
+    decoder
 end # struct
 
 # Mark function as Flux.Functors.@functor so that Flux.jl allows for training
 Flux.@functor Decoder
+
+# ------------------------------------------------------------------------------
+
+# Custom display method for the type
+function Base.show(io::IO, ::Type{Decoder{D}}) where {D}
+    print(io, "Decoder{…}")
+end
+
+# ------------------------------------------------------------------------------
 
 @doc raw"""
     Decoder(n_input, n_latent, decoder_neurons, decoder_activation, 
@@ -269,7 +280,7 @@ end # function
 # ==============================================================================
 
 @doc raw"""
-    SimpleGaussianDecoder <: AbstractGaussianDecoder
+    SimpleGaussianDecoder{D} <: AbstractGaussianDecoder
 
 A straightforward decoder structure for variational autoencoders (VAEs) that
 contains only a single decoder network.
@@ -285,16 +296,19 @@ It's commonly used when the VAE's latent space distribution is implicitly
 defined, and there's no need for separate paths or operations on the mean or log
 standard deviation.
 """
-struct SimpleGaussianDecoder <: AbstractGaussianDecoder
-    decoder::Flux.Chain
+@concrete struct SimpleGaussianDecoder <: AbstractGaussianDecoder
+    decoder
 end # struct
 
 # Mark function as Flux.Functors.@functor so that Flux.jl allows for training
 Flux.@functor SimpleGaussianDecoder
 
 @doc raw"""
-    SimpleGaussianDecoder(n_input, n_latent, decoder_neurons, decoder_activation, 
-                output_activation; init=Flux.glorot_uniform)
+    SimpleGaussianDecoder(
+        n_input, n_latent, decoder_neurons, 
+        decoder_activation, output_activation; 
+        init=Flux.glorot_uniform
+    )
 
 Constructs and initializes a `SimpleGaussianDecoder` object designed for
 variational autoencoders (VAEs). This function sets up a straightforward decoder
@@ -430,7 +444,7 @@ end # function
 # ==============================================================================
 
 @doc raw"""
-    JointGaussianLogDecoder <: AbstractGaussianLogDecoder
+    JointGaussianLogDecoder{D<:Flux.Chain,L<:Flux.Dense} <: AbstractGaussianLogDecoder
 
 An extended decoder structure for VAEs that incorporates separate layers for
 mapping from the latent space to both its mean (`µ`) and log standard deviation
@@ -449,14 +463,23 @@ mapping from the latent space to both its mean (`µ`) and log standard deviation
 decoder network is used initially, and then splits into two separate paths for
[…]
