### 2024-07-07 fcbe8d3
update type checking tests with new concrete structs
