### 2024-07-08 mrazomej/AutoEncode.jl#14 @fabian-s
Thank you both!
