### 2024-07-08 dfbc76f
add typeof function for SimpleGaussianDecoder
