### 2024-07-08 f3794e0
add tests for GPU implementation
