+++ test/vae.jl
@@ -7,6 +7,8 @@ import AutoEncoderToolkit.VAEs
 import AutoEncoderToolkit.regularization
 # Import Flux library
 import Flux
+# Import CUDA
+using CUDA
 
 # Import basic math
 import Random
@@ -351,7 +353,7 @@ end # @testset "VAE gradient"
     # Define batch of data
     x_matrix = data
 
-    @testset "without regularization" begin
+    @testset "CPU | without regularization" begin
         # Loop through decoders
         for decoder in decoders
             # Define VAE with any decoder
@@ -382,61 +384,44 @@ end # @testset "VAE gradient"
             # Extract modified parameters
             # params_end = deepcopy(Flux.params(vae))
         end # for decoder in decoders
-    end # @testset "without regularization"
+    end # @testset "CPU | without regularization"
 
-    # @testset "with regularization" begin
-    #     reg_function = regularization.l2_regularization
-    #     reg_kwargs = Dict(:reg_terms => [:encoder_μ, :encoder_logσ])
-    #     # Loop through decoders
-    #     for decoder in decoders
-    #         # Define VAE with any decoder
-    #         vae = deepcopy(joint_log_encoder) * decoder
+    if CUDA.functional()
+        @testset "GPU | without regularization" begin
+            # Loop through decoders
+            for decoder in decoders
+                # Define VAE with any decoder
+                vae = deepcopy(joint_log_encoder) *
+                      deepcopy(decoder)
+                # Upload to GPU
+                vae = Flux.gpu(vae)
 
-    #         # Explicit setup of optimizer
-    #         opt_state = Flux.Train.setup(
-    #             Flux.Optimisers.Adam(1E-3),
-    #             vae
-    #         )
+                # Explicit setup of optimizer
+                opt_state = Flux.Train.setup(
+                    Flux.Optimisers.Adam(1E-2),
+                    vae
+                )
 
-    #         # Extract parameters
-    #         params_init = deepcopy(Flux.params(vae))
+                # Loop through a couple of epochs
+                losses = Float32[]  # Track the loss
+                for epoch = 1:n_epochs
+                    Random.seed!(42)
+                    # Test training function
+                    L = VAEs.train!(
+                        vae, CUDA.cu(data), opt_state; loss_return=true
+                    )
+                    push!(losses, L)
+                end
 
-    #         # Loop through a couple of epochs
-    #         losses = Float32[]  # Track the loss
-    #         for epoch = 1:n_epochs
-    #             Random.seed!(42)
-    #             # Test training function
-    #             VAEs.train!(
-    #                 vae, data, opt_state;
-    #                 loss_kwargs=Dict(
-    #                     :reg_function => reg_function,
-    #                     :reg_kwargs => reg_kwargs
-    #                 )
-    #             )
-    #             push!(
-    #                 losses,
-    #                 VAEs.loss(
-    #                     vae, data;
-    #                     reg_function=reg_function, reg_kwargs=reg_kwargs
-    #                 )
-    #             )
-    #         end
+                # Check if loss is decreasing
+                @test all(diff(losses) ≠ 0)
 
-    #         # Check if loss is decreasing
-    #         @test all(diff(losses) ≠ 0)
+                # Extract modified parameters
+                # params_end = deepcopy(Flux.params(vae))
+            end # for decoder in decoders
+        end # @testset "GPU | without regularization"
+    end # if CUDA.functional()
 
-    #         # Extract modified parameters
-    #         params_end = deepcopy(Flux.params(vae))
-
-    #         # Check that parameters have significantly changed
-    #         threshold = 1e-5
-    #         # Check if any parameter has changed significantly
-    #         @test all([
[…]
