+++ test/rhvae.jl
@@ -8,6 +8,8 @@ import AutoEncoderToolkit.VAEs
 import AutoEncoderToolkit.utils
 # Import Flux library
 import Flux
+# Import GPU library
+using CUDA
 
 # Import basic math
 import LinearAlgebra
@@ -1562,13 +1564,8 @@ end # @testset "HVAE grad"
 
 ## =============================================================================
 
-# NOTE: The following tests are commented out because they fail with GitHub
-# Actions with the following error:
-# Got exception outside of a @test
-# BoundsError: attempt to access 16-element Vector{UInt8} at index [0]
-
 @testset "RHVAE training" begin
-    @testset "without regularization" begin
+    @testset "CPU | without regularization" begin
         # Loop through decoders
         for decoder in decoders
             # Build temporary RHVAE
@@ -1591,6 +1588,34 @@ end # @testset "HVAE grad"
             @test isa(L, Float32)
         end # for decoder in decoders
     end # @testset "without regularization"
+
+    if CUDA.functional()
+        @testset "GPU | without regularization" begin
+            # Loop through decoders
+            for decoder in decoders
+                # Build temporary RHVAE
+                rhvae = RHVAEs.RHVAE(
+                    joint_log_encoder * decoder,
+                    metric_chain,
+                    data,
+                    T,
+                    λ,
+                ) |> Flux.gpu
+
+                # Explicit setup of optimizer
+                opt_state = Flux.Train.setup(
+                    Flux.Optimisers.Adam(1E-2),
+                    rhvae
+                )
+
+                # Test training function
+                L = RHVAEs.train!(
+                    rhvae, CUDA.cu(data), opt_state; loss_return=true
+                )
+                @test isa(L, Float32)
+            end # for decoder in decoders
+        end # @testset "without regularization"
+    end # if CUDA.functional()
 end # @testset "RHVAE training"
 
 println("\nAll tests passed!\n")
