### 2024-07-08 199ef64
add GPU-compatible functions to train MMDVAEs
