+++ ext/AutoEncoderToolkitCUDAExt/mmdvae.jl
@@ -1,0 +1,358 @@
+# Import basic math
+import StatsBase
+
+import AutoEncoderToolkit as AET
+using AutoEncoderToolkit.MMDVAEs
+
+# Import AutoDiff backends
+using ChainRulesCore: ignore_derivatives
+
+@doc raw"""
+    gaussian_kernel(
+        x::CuArray, y::CuArray; ρ::Float32=1.0f0
+    )
+GPU-compatible function to compute the Gaussian Kernel between columns of arrays `x` and `y`, defined
+as 
+        k(x_i, y_j) = exp(-||x_i - y_j||² / ρ²)
+# Arguments
+- `x::CuArray`: First input array for the kernel.
+- `y::CuArray`: Second input array for the kernel.  
+# Optional Keyword Arguments
+- `ρ=1.0f0`: Kernel amplitude hyperparameter. Larger ρ gives a smoother
+  kernel.
+# Returns
+- `k::CuArray`: Kernel matrix where each element (i,j) is computed as k(x_i, y_j)
+# Theory
+The Gaussian kernel measures the similarity between two points `x_i` and `y_j`. It
+is widely used in many machine learning algorithms. This implementation computes
+the squared Euclidean distance between all pairs of columns in `x` and `y`, scales
+the distance by ρ² and takes the exponential.
+"""
+function MMDVAEs.gaussian_kernel(
+    x::CUDA.CuArray,
+    y::CUDA.CuArray;
+    ρ::Float32=1.0f0
+)
+    # Compute dot products
+    xx = sum(abs2, x, dims=1)
+    yy = sum(abs2, y, dims=1)
+    xy = CUDA.transpose(x) * y
+
+    # Compute pairwise squared Euclidean distances
+    dist = CUDA.transpose(xx) .+ yy .- 2 .* xy
+
+    # Scale distances and compute exponential
+    return exp.(-dist ./ (ρ^2 * size(x, 1)))
+end
+
+# ------------------------------------------------------------------------------
+
+@doc raw"""
+    loss(mmdvae::MMDVAE, x::AbstractArray; σ::Number=1.0f0, λ::Number=1.0f0, α::Number=0.0f0, n_latent_samples::Int=50, kernel::Function=gaussian_kernel, kernel_kwargs::Union{NamedTuple,Dict}=Dict(), reconstruction_loglikelihood::Function=decoder_loglikelihood, kl_divergence::Function=encoder_kl)
+
+Loss function for the Maximum-Mean Discrepancy variational autoencoder
+(MMD-VAE). The loss function is defined as:
+
+loss = -⟨log p(x|z)⟩ + (1 - α) * Dₖₗ(qᵩ(z | x) || p(z)) + (λ + α - 1) *
+MMD-D(qᵩ(z) || p(z)),
+
+# Arguments
+- `mmdvae::MMDVAE`: Struct containing the elements of the MMD-VAE.
+- `x::AbstractArray`: Input data.
+
+# Optional Arguments
+- `λ::Number=1.0f0`: Hyperparameter that emphasizes the importance of the KL
+  divergence between qᵩ(z) and π(z) during training.
+- `α::Number=0.0f0`: Hyperparameter that emphasizes the importance of the Mutual
+  Information term during optimization.
+- `n_latent_samples::Int=50`: Number of samples to take from the latent space
+  prior π(z) when computing the MMD divergence.
+- `kernel::Function=gaussian_kernel`: Kernel used to compute the divergence.
+  Default is the Gaussian Kernel.
+- `kernel_kwargs::Union{NamedTuple,Dict}=Dict()`: Additional keyword arguments
+  to be passed to the kernel function.
[…]
