### 2024-07-08 77c6efd
update MMDVAEs test to include GPU-compatible functions
