### 2024-07-08 43dbd96
small fix to upload output of finite difference to GPU
