### 2024-07-08 e9c780d
update tests to include GPU training
