### 2024-07-08 8b80d92
add tests for CUDA extension-specific functions
