+++ src/simplechains.jl
@@ -1,232 +1,0 @@
-@doc raw"""
-    SimpleAE
-
-`mutable struct` representing a basic autoencoder (AE) specifically designed for
-the `SimpleChains.jl` package.
-
-# Fields
-- `AE::SimpleChains.SimpleChain`: The autoencoder model defined as a sequence of
-  layers and operations using the `SimpleChain` construct from the
-  `SimpleChains.jl` package. This chain encompasses both the encoder and decoder
-  parts of the AE.
-- `param::DenseArray{Float32}`: An array of model parameters, presumably to aid
-  in training or other operations. The exact nature and organization of these
-  parameters depend on the AE's design and requirements.
-
-# Description
-This struct encapsulates the components of a basic autoencoder model suitable
-for use with the `SimpleChains.jl` framework. While the autoencoder's
-architecture is defined in the `AE` field, associated parameters, possibly used
-for optimization routines or other tasks, are stored in the `param` field.
-
-# Note
-Users should ensure compatibility between the `SimpleAE` structure and
-functions/methods provided by `SimpleChains.jl` when integrating or extending
-this structure.
-"""
-mutable struct SimpleAE <: AbstractDeterministicAutoEncoder
-    AE::SimpleChains.SimpleChain
-    param::DenseArray{Float32}
-end # struct
-
-
-##
-
-@doc raw"""
-    SimpleAE(
-        n_input::Int, 
-        n_latent::Int, 
-        latent_activation::Function,
-        output_activation::Function,
-        encoder::Vector{Int}, 
-        encoder_activation::Vector{Function},
-        decoder::Vector{Int}, 
-        decoder_activation::Vector{Function}
-    ) -> SimpleAE
-
-Constructs and initializes a `SimpleAE` autoencoder using components from the
-`SimpleChains.jl` package.
-
-# Arguments
-- `n_input::Int`: Dimensionality of the input space.
-- `n_latent::Int`: Dimensionality of the latent space.
-- `latent_activation::Function`: Activation function applied at the latent space
-  layer.
-- `output_activation::Function`: Activation function applied at the output
-  layer.
-- `encoder::Vector{Int}`: Specifies the dimensions of each hidden layer in the
-  encoder network.
-- `encoder_activation::Vector{Function}`: Activation functions for each hidden
-  layer in the encoder. The number of functions should match the length of
-  `encoder`.
-- `decoder::Vector{Int}`: Specifies the dimensions of each hidden layer in the
-  decoder network.
-- `decoder_activation::Vector{Function}`: Activation functions for each hidden
-  layer in the decoder. The number of functions should match the length of
-  `decoder`.
-
-# Returns
-- A `SimpleAE` instance representing the constructed autoencoder.
-
-# Description
-The function creates a `SimpleAE` autoencoder by sequentially combining the
-provided encoder and decoder layers, while ensuring the correct matching of
-activation functions for each layer. The resultant autoencoder is structured to
-operate seamlessly within the `SimpleChains.jl` ecosystem.
-
-# Example
-```julia
-ae_model = SimpleAE(
-    784, 32, relu, sigmoid, [512, 256], [relu, relu], [256, 512], [relu, relu]
-)
-```
-
-# Note
-Ensure that the number of dimensions in encoder matches the number of activation
-functions in encoder_activation, and similarly for decoder and
-decoder_activation.
-"""
-function SimpleAE(
-    n_input::Int,
-    n_latent::Int,
-    latent_activation::Function,
-    output_activation::Function,
-    encoder::Vector{<:Int},
-    encoder_activation::Vector{<:Function},
-    decoder::Vector{<:Int},
-    decoder_activation::Vector{<:Function}
-)
-    # Check there's enough activation functions for all layers
[…]
