+++ src/regularization.jl
@@ -1,191 +1,0 @@
-# Import Abstract Types
-using ..AutoEncoderToolkit: JointGaussianLogEncoder, SimpleGaussianDecoder, JointGaussianLogDecoder, SplitGaussianLogDecoder
-
-# ==============================================================================
-
-@doc raw"""
-        l2_regularization(outputs::NamedTuple, reg_terms::Vector{Symbol})
-
-Compute the L2 regularization term (also known as Ridge regularization) based on
-the autoencoder outputs and the `reg_terms` vector.
-
-L2 regularization is defined as: L₂(v) = λ ∑ᵢ vᵢ²
-
-where:
-- λ is the regularization strength (not computed in this function, but typically
-  applied outside).
-- vᵢ represents each element of the vector or matrix v.
-
-The primary purpose of L2 regularization is to add a penalty to the magnitude of
-the model parameters. By doing so, it discourages the model from fitting the
-training data too closely (overfitting) and promotes smoother decision
-boundaries or function approximations. This helps in improving generalization to
-new, unseen data.
-
-# Arguments
-- `outputs::NamedTuple`: NamedTuple containing outputs from the AE or VAE model.
-
-## Optional Keyword Arguments
-- `reg_terms::Vector{Symbol}`: Vector that specifies which entries from
-  `outputs` to consider for regularization.
-
-# Returns
-- `reg_term::Float32`: The computed L2 regularization value.
-
-# Notes
-- Ensure that all elements in `reg_terms` are keys in `outputs`.
-"""
-function l2_regularization(
-    outputs::NamedTuple; reg_terms::Vector{Symbol}=[:decoder_µ]
-)::Float32
-    # Ensyre there is at least one key
-    if isempty(reg_terms)
-        return 0.0f0
-    end # if
-
-    # Ensure all keys in reg_terms are in outputs
-    if !all(key ∈ keys(outputs) for key in reg_terms)
-        error("All keys in reg_terms must exist in outputs!")
-    end # if
-
-    # Compute the regularization term without in-place mutation
-    reg_term = sum(sum(outputs[term] .^ 2) for term in reg_terms)
-
-    return reg_term
-end # function
-
-# ==============================================================================
-
-@doc raw"""
-        min_variance_regularization(outputs::NamedTuple, σ_min::Float32=0.1f0, logσ::Bool=true)
-
-Compute the minimum variance constraint regularization term based on the VAE
-outputs and a specified minimum standard deviation (σ_min).
-
-The regularization is defined as: 
-- If `logσ` is true: L = λ ∑ᵢ max(0, log(σᵢ²) - log(σ_min²))
-- If `logσ` is false: L = λ ∑ᵢ max(0, σᵢ² - σ_min²)
-
-Where:
-- λ is the regularization strength (not computed in this function, but typically
-  applied outside).
-- σᵢ represents the standard deviation for each output in the VAE decoder.
-
-The primary purpose of this regularization is to prevent the variance from going
-below a certain threshold, discouraging the model from making overconfident
-predictions.
-
-# Arguments
-- `outputs::NamedTuple`: NamedTuple containing outputs from the VAE model.
-
-## Optional Keyword Arguments
-- `σ_min::Float32`: The minimum allowable standard deviation. Default is 0.1.
-- `logσ::Bool`: If true, the regularization is computed in the log-space. If
-  false, the regularization is computed in the original space. Default is true.
-
-# Returns
-- `reg_term::Float32`: The computed minimum variance constraint regularization
-  value.
-"""
-function min_variance_regularization(
-    outputs::NamedTuple; σ_min::Float32=0.1f0, logσ::Bool=true
-)::Float32
[…]

[…]
