### 2024-07-08 mrazomej/AutoEncode.jl#14 @mrazomej
@avik-pal and @fabian-s

I implemented the requested changes and released the [`v0.1.0`](https://github.com/mrazomej/AutoEncoderToolkit.jl/releases/tag/v0.1.0) version. The coverage reported does not reflect the package's real coverage. I accidentally kept a lot of "experimental" files in the main branch, but those have been removed. Basically, every function in the available models has been fully tested.

Thanks again, @avik-pal, for the incredibly insightful comments. Your suggestions truly improve the quality of the package.

Please let me know if there is anything else you would like me to implement. If not, we can close this issue.
