### 2024-07-10 #6794 @fabian-s
@avik-pal do these latest updates  remedy the remaining issues re functionality, automated tests and contributing guidelines in your assessement?
