### 2024-07-10 #6794 @sandeshkatakam
> @sandeshkatakam i take it you found no issues at all and are entirely happy with the package as is, no comments or remarks?
>

I raised an issue with automated tests for the package and the author responded to me a while ago. Yeah, overall no issues from my side 👍
