### 2024-07-11 mrazomej/AutoEncode.jl#14 @avik-pal
The CI isn't running correctly right? It proceeds on error https://github.com/mrazomej/AutoEncoderToolkit.jl/blob/8b80d925ef9a50f928fe746b3195e914730e91ca/.github/workflows/CI.yml#L50 https://github.com/mrazomej/AutoEncoderToolkit.jl/actions/runs/9846125121/job/27183232171#step:7:281
