+++ test/regularization.jl
@@ -1,293 +1,0 @@
-##
-println("Testing regularization module:\n")
-##
-
-# Import AutoEncoderToolkit.jl module to be tested
-import AutoEncoderToolkit.regularization
-import AutoEncoderToolkit.VAEs
-# Import Flux library
-import Flux
-
-# Import basic math
-import Random
-import StatsBase
-import Distributions
-
-Random.seed!(42)
-
-## =============================================================================
-
-# Define number of inputs
-n_input = 3
-# Define number of synthetic data points
-n_data = 5
-# Define number of hidden layers
-n_hidden = 2
-# Define number of neurons in non-linear hidden layers
-n_neuron = 3
-# Define dimensionality of latent space
-n_latent = 2
-
-## =============================================================================
-println("Generating synthetic data...")
-
-# Define function
-f(x₁, x₂) = 10 * exp(-(x₁^2 + x₂^2))
-
-# Defien radius
-radius = 3
-
-# Sample random radius
-r_rand = radius .* sqrt.(Random.rand(n_data))
-
-# Sample random angles
-θ_rand = 2π .* Random.rand(n_data)
-
-# Convert form polar to cartesian coordinates
-x_rand = Float32.(r_rand .* cos.(θ_rand))
-y_rand = Float32.(r_rand .* sin.(θ_rand))
-# Feed numbers to function
-z_rand = f.(x_rand, y_rand)
-
-# Compile data into matrix
-data = Matrix(hcat(x_rand, y_rand, z_rand)')
-
-## =============================================================================
-
-println("Defining layers structure...")
-
-# Define latent space activation function
-latent_activation = Flux.identity
-# Define output layer activation function
-output_activation = Flux.identity
-# Define output layer activation function
-output_activations = [Flux.identity, Flux.softplus]
-
-# Define encoder layer and activation functions
-encoder_neurons = repeat([n_neuron], n_hidden)
-encoder_activation = repeat([Flux.swish], n_hidden)
-
-# Define decoder layer and activation function
-decoder_neurons = repeat([n_neuron], n_hidden)
-decoder_activation = repeat([Flux.swish], n_hidden)
-
-# Define decoder split layers and activation functions
-µ_neurons = repeat([n_neuron], n_hidden)
-µ_activation = repeat([Flux.swish], n_hidden)
-
-logσ_neurons = repeat([n_neuron], n_hidden)
-logσ_activation = repeat([Flux.swish], n_hidden)
-
-σ_neurons = repeat([n_neuron], n_hidden)
-σ_activation = [repeat([Flux.swish], n_hidden - 1); Flux.softplus]
-
-
-## =============================================================================
-
-
-println("Defining encoders...")
-# Initialize JointGaussianLogEncoder
-joint_log_encoder = VAEs.JointGaussianLogEncoder(
-    n_input,
-    n_latent,
-    encoder_neurons,
-    encoder_activation,
-    latent_activation,
-)
-
-# -----------------------------------------------------------------------------
-
-println("Defining decoders...")
-
-# Initialize SimpleGaussianDecoder
-simple_decoder = VAEs.SimpleGaussianDecoder(
-    n_input,
-    n_latent,
-    decoder_neurons,
-    decoder_activation,
-    output_activation
-)
-
-# Initialize JointGaussianLogDecoder
-joint_log_decoder = VAEs.JointGaussianLogDecoder(
-    n_input,
-    n_latent,
-    decoder_neurons,
-    decoder_activation,
-    output_activation
-)
-
-# Initialize SplitGaussianLogDecoder
-split_log_decoder = VAEs.SplitGaussianLogDecoder(
-    n_input,
-    n_latent,
-    µ_neurons,
-    µ_activation,
-    logσ_neurons,
-    logσ_activation,
-)
-
-# Initialize JointGaussianDecoder
[…]
