### 2024-07-11 5e99584
remove call to regularization
