### 2024-07-11 7a7302f
update which modules to add
