### 2024-07-11 e6d41d5
fix checking if CUDA is available for testing
