+++ test/hvae.jl
@@ -8,14 +8,16 @@ import AutoEncoderToolkit.VAEs
 import AutoEncoderToolkit.utils
 # Import Flux library
 import Flux
-# Import CUDA
-using CUDA
 
 # Import basic math
 import LinearAlgebra
 import Random
 import StatsBase
 import Distributions
+
+using Pkg: project
+# Check if CUDA is available
+cuda_functional = haskey(project().dependencies, "CUDA")
 
 Random.seed!(42)
 
@@ -800,7 +802,8 @@ end # @testset "HVAE grad"
         end # for decoder in decoders
     end # @testset "CPU | without regularization"
 
-    if CUDA.functional()
+    if cuda_functional
+        using CUDA
         @testset "GPU | without regularization" begin
             # Loop through decoders
             for decoder in decoders
@@ -822,7 +825,7 @@ end # @testset "HVAE grad"
                 @test isa(L, Float32)
             end # for decoder in decoders
         end # @testset "CPU | without regularization"
-    end # if CUDA.functional()
+    end # if cuda_functional
 end # @testset "HVAE training"
 
 println("\nAll tests passed!\n")
