### 2024-07-11 2f9702d
change continue-on-error
