+++ .github/workflows/CI.yml
@@ -47,17 +47,17 @@ jobs:
       - name: "Run test without coverage report"
         uses: julia-actions/julia-runtest@v1
         if: ${{ !contains(fromJson('["1", "1.9"]'), matrix.version) || matrix.os != 'ubuntu-latest' }}
-        continue-on-error: true
+        continue-on-error: false
         with:
           coverage: false
 
       - name: "Run test with coverage report"
         uses: julia-actions/julia-runtest@v1
         if: contains(fromJson('["1", "1.9"]'), matrix.version) && matrix.os == 'ubuntu-latest'
-        continue-on-error: true
+        continue-on-error: false
       - uses: julia-actions/julia-processcoverage@v1
         if: contains(fromJson('["1", "1.9"]'), matrix.version) && matrix.os == 'ubuntu-latest'
-        continue-on-error: true
+        continue-on-error: false
       - uses: codecov/codecov-action@v4.5.0
         if: contains(fromJson('["1", "1.9"]'), matrix.version) && matrix.os == 'ubuntu-latest'
         with:
