+++ test/mmdvae.jl
@@ -12,9 +12,9 @@ import StatsBase
 import Distributions
 import Distances
 
-using Pkg: project
+import Pkg
 # Check if CUDA is available
-cuda_functional = haskey(project().dependencies, "CUDA")
+cuda_functional = haskey(Pkg.project().dependencies, "CUDA")
 
 if cuda_functional
     println("CUDA available - running tests with CUDA.")
