+++ test/cuda_ext.jl
@@ -2,9 +2,9 @@ using AutoEncoderToolkit
 using Test
 using Flux
 
-using Pkg: project
+import Pkg
 # Check if CUDA is available
-cuda_functional = haskey(project().dependencies, "CUDA")
+cuda_functional = haskey(Pkg.project().dependencies, "CUDA")
 
 if cuda_functional
     println("CUDA available - running tests with CUDA.")
