### 2024-07-11 mrazomej/AutoEncode.jl#14 @mrazomej
> The CI isn't running correctly right? It proceeds on error
>
> https://github.com/mrazomej/AutoEncoderToolkit.jl/blob/8b80d925ef9a50f928fe746b3195e914730e91ca/.github/workflows/CI.yml#L50
>
> https://github.com/mrazomej/AutoEncoderToolkit.jl/actions/runs/9846125121/job/27183232171#step:7:281

@avik-pal , you are absolutely right, and I apologize for that. I made an error in determining whether or not to test the `CUDA`- related functions on GitHub, but now that is fixed. I changed that `continue-on-error` to `false`.

However, I'm running into an issue I don't really understand. Locally, all tests pass whether `CUDA` is available or not.
```julia
Test Summary:         | Pass  Total     Time
AutoEncoderToolkit.jl | 1387   1387  1m44.3s
```
But on GitHub, there seems to be a recurrent issue when taking gradients with `Zygote` (see example below from [here](https://github.com/mrazomej/AutoEncoderToolkit.jl/actions/runs/9897780064/job/27343079815)). This causes all functions where a gradient is taken to fail.
```julia
Test Summary:                               | Pass  Error  Total     Time
AutoEncoderToolkit.jl                       | 1344     30   1374  2m54.0s
```
That was the original reason I had set `continue-on-error: true`. I am not sure how to fix this issue.

```julia
NeuralGeodesic training: Error During Test at /Users/runner/work/AutoEncoderToolkit.jl/AutoEncoderToolkit.jl/test/diffgeo.jl:300
  Got exception outside of a @test
  BoundsError: attempt to access 16-element Vector{UInt8} at index [0]
  Stacktrace:
    [1] getindex
      @ ./essentials.jl:13 [inlined]
    [2] isslotfilled
      @ ./dict.jl:162 [inlined]
    [3] skip_deleted
      @ ./dict.jl:706 [inlined]
    [4] iterate
      @ ./dict.jl:724 [inlined]
    [5] _pullback
      @ ~/.julia/packages/Zygote/nsBv0/src/lib/base.jl:53 [inlined]
    [6] _pullback
      @ ~/.julia/packages/Zygote/nsBv0/src/lib/base.jl:70 [inlined]
    [7] merge
      @ ./namedtuple.jl:361 [inlined]
    [8] _pullback(::Zygote.Context{false}, ::typeof(merge), ::@NamedTuple{}, ::Dict{Any, Any})
      @ Zygote ~/.julia/packages/Zygote/nsBv0/src/compiler/interface2.jl:0
    [9] #11
      @ ~/work/AutoEncoderToolkit.jl/AutoEncoderToolkit.jl/src/diffgeo/neural_geodesics.jl:562 [inlined]
   [10] _pullback(ctx::Zygote.Context{false}, f::AutoEncoderToolkit.diffgeo.NeuralGeodesics.var"#11#12"{typeof(AutoEncoderToolkit.diffgeo.NeuralGeodesics.loss), Dict{Any, Any}, RHVAE{AutoEncoderToolkit.VAEs.VAE{JointGaussianLogEncoder{…}, SimpleGaussianDecoder{…}}}, Vector{Float32}}, args::AutoEncoderToolkit.diffgeo.NeuralGeodesics.NeuralGeodesic)
      @ Zygote ~/.julia/packages/Zygote/nsBv0/src/compiler/interface2.jl:0
   [11] pullback(f::Function, cx::Zygote.Context{false}, args::AutoEncoderToolkit.diffgeo.NeuralGeodesics.NeuralGeodesic)
      @ Zygote ~/.julia/packages/Zygote/nsBv0/src/compiler/interface.jl:90
   [12] pullback
      @ ~/.julia/packages/Zygote/nsBv0/src/compiler/interface.jl:88 [inlined]
   [13] withgradient(f::Function, args::AutoEncoderToolkit.diffgeo.NeuralGeodesics.NeuralGeodesic)
      @ Zygote ~/.julia/packages/Zygote/nsBv0/src/compiler/interface.jl:205
   [14] train!(curve::AutoEncoderToolkit.diffgeo.NeuralGeodesics.NeuralGeodesic, rhvae::RHVAE{AutoEncoderToolkit.VAEs.VAE{JointGaussianLogEncoder{…}, SimpleGaussianDecoder{…}}}, t::Vector{Float32}, opt::@NamedTuple{mlp::@NamedTuple{layers::Tuple{@NamedTuple{weight::Optimisers.Leaf{Optimisers.Adam, Tuple{Matrix{Float32}, Matrix{Float32}, Tuple{Float32, Float32}}}, bias::Optimisers.Leaf{Optimisers.Adam, Tuple{Vector{Float32}, Vector{Float32}, Tuple{Float32, Float32}}}, σ::Tuple{}}, @NamedTuple{weight::Optimisers.Leaf{Optimisers.Adam, Tuple{Matrix{Float32}, Matrix{Float32}, Tuple{Float32, Float32}}}, bias::Optimisers.Leaf{Optimisers.Adam, Tuple{Vector{Float32}, Vector{Float32}, Tuple{Float32, Float32}}}, σ::Tuple{}}, @NamedTuple{weight::Optimisers.Leaf{Optimisers.Adam, Tuple{Matrix{Float32}, Matrix{Float32}, Tuple{Float32, Float32}}}, bias::Optimisers.Leaf{Optimisers.Adam, Tuple{Vector{Float32}, Vector{Float32}, Tuple{Float32, Float32}}}, σ::Tuple{}}}}}; loss_function::typeof(AutoEncoderToolkit.diffgeo.NeuralGeodesics.loss), loss_kwargs::Dict{Any, Any}, verbose::Bool, loss_return::Bool)
      @ AutoEncoderToolkit.diffgeo.NeuralGeodesics ~/work/AutoEncoderToolkit.jl/AutoEncoderToolkit.jl/src/diffgeo/neural_geodesics.jl:561
   [15] macro expansion
      @ ~/work/AutoEncoderToolkit.jl/AutoEncoderToolkit.jl/test/diffgeo.jl:324 [inlined]
   [16] macro expansion
      @ ~/hostedtoolcache/julia/1.10.4/x64/share/julia/stdlib/v1.10/Test/src/Test.jl:1577 [inlined]
   [17] top-level scope
      @ ~/work/AutoEncoderToolkit.jl/AutoEncoderToolkit.jl/test/diffgeo.jl:302
   [18] include(fname::String)
      @ Base.MainInclude ./client.jl:489
   [19] macro expansion
      @ ~/work/AutoEncoderToolkit.jl/AutoEncoderToolkit.jl/test/runtests.jl:45 [inlined]
   [20] macro expansion
      @ ~/hostedtoolcache/julia/1.10.4/x64/share/julia/stdlib/v1.10/Test/src/Test.jl:1577 [inlined]
   [21] top-level scope
      @ ~/work/AutoEncoderToolkit.jl/AutoEncoderToolkit.jl/test/runtests.jl:18
   [22] include(fname::String)
      @ Base.MainInclude ./client.jl:489
   [23] top-level scope
      @ none:6
   [24] eval
      @ ./boot.jl:385 [inlined]
   [25] exec_options(opts::Base.JLOptions)
      @ Base ./client.jl:291
   [26] _start()
      @ Base ./client.jl:552
  ```
