### 2024-07-13 mrazomej/AutoEncode.jl#14 @avik-pal
Not sure why there is a discrepancy from local runs but doing a merge of `Dict` and `NamedTuple` doesn't seem to have a rrule. Regardless you should do a merge of 2 NamedTuples, `Dict`s generally have unnecessary overheads.
