### 2024-07-15 29bc4f3
replace all dictionaries with namedtuples
