+++ src/hvae.jl
@@ -508,7 +508,7 @@ const ∇energyhvae = (
         reconstruction_loglikelihood::Function=decoder_loglikelihood,
         latent_logprior::Function=spherical_logprior,
         adtype::Union{Symbol,Nothing}=nothing,
-        adkwargs::Union{NamedTuple,Dict}=Dict(),
+        adkwargs::NamedTuple=NamedTuple(),
     )
 
 Compute the gradient of the potential energy of a Hamiltonian Variational
@@ -538,7 +538,7 @@ potential energy computed for given data `x` and latent variable `z`.
   is `spherical_logprior`.  
 - `adtype::Symbol`=:finite`: The type of automatic differentiation method to
   use. Must be `:finite` or `:TaylorDiff`. Default is `:finite`.
-- `adkwargs::Union{NamedTuple,Dict}=Dict()`: Additional keyword arguments to
+- `adkwargs::NamedTuple=NamedTuple()`: Additional keyword arguments to
   pass to the automatic differentiation method.
 
 # Returns
@@ -553,7 +553,7 @@ function ∇potential_energy(
     reconstruction_loglikelihood::Function=decoder_loglikelihood,
     latent_logprior::Function=spherical_logprior,
     adtype::Symbol=:finite,
-    adkwargs::Union{NamedTuple,Dict}=Dict(),
+    adkwargs::NamedTuple=NamedTuple(),
 )
     # Check that the provided adtype is valid
     if adtype ∉ keys(∇energyhvae)
@@ -580,7 +580,7 @@ end # function
         reconstruction_loglikelihood::Function=decoder_loglikelihood,
         latent_logprior::Function=spherical_logprior,
         adtype::Union{Symbol,Nothing}=nothing,
-        adkwargs::Union{NamedTuple,Dict}=Dict(),
+        adkwargs::NamedTuple=NamedTuple(),
     )
 
 Compute the gradient of the potential energy of a Hamiltonian Variational
@@ -608,7 +608,7 @@ potential energy computed for given data `x` and latent variable `z`.
   is `spherical_logprior`.  
   - `adtype::Symbol`=:finite`: The type of automatic differentiation method to
   use. Must be `:finite` or `:TaylorDiff`. Default is `:finite`.
-- `adkwargs::Union{NamedTuple,Dict}=Dict()`: Additional keyword arguments to
+- `adkwargs::NamedTuple=NamedTuple()`: Additional keyword arguments to
   pass to the automatic differentiation method.
 
 # Returns
@@ -622,7 +622,7 @@ function ∇potential_energy(
     reconstruction_loglikelihood::Function=decoder_loglikelihood,
     latent_logprior::Function=spherical_logprior,
     adtype::Symbol=:finite,
-    adkwargs::Union{NamedTuple,Dict}=Dict(),
+    adkwargs::NamedTuple=NamedTuple(),
 )
     # Check that the provided adtype is valid
     if adtype ∉ keys(∇energyhvae)
@@ -651,7 +651,7 @@ end # function
         decoder::AbstractVariationalDecoder,
         decoder_output::NamedTuple;
         ϵ::Union{<:Number,<:AbstractVector}=Float32(1E-4),
-        ∇U_kwargs::Union{Dict,NamedTuple}=(
+        ∇U_kwargs::NamedTuple=(
             reconstruction_loglikelihood=reconstruction_loglikelihood,
             latent_logprior=spherical_logprior,
         )
@@ -690,7 +690,7 @@ This function performs these three steps in sequence.
[…]

[…]
