### 2024-07-15 mrazomej/AutoEncode.jl#14 @mrazomej
@avik-pal , that was it! Replacing all `Dict` with `NamedTuple` solved the issue. Thank you so much for your help; now all tests pass on `GitHub.`
```julia
All tests passed!

Test Summary:         | Pass  Total     Time
AutoEncoderToolkit.jl | 1387   1387  8m08.4s
     Testing AutoEncoderToolkit tests passed
```
I released and registered `v0.1.1` with all these changes. Thanks again for your help!
Please let me know if you would like me to implement anything else.
