### 2024-07-16 #6794 @fabian-s
@avik-pal thanks also from my side for your  tremendous input on the package code.  This is the way peer review should work.
In your opinion, do the latest changes remedy the remaining issues re functionality, automated tests and community guidelines?
