### 2024-07-19 mrazomej/AutoEncode.jl#14 @avik-pal
Things look great now.
