### 2024-07-23 e3ae693
make MetricChain type stable. Define v0.1.2
