### 2024-07-23 5fe0af7
add compat for ConcreteStructs
