+++ paper.md
@@ -25,75 +25,74 @@ With the advent of generative models, the field of unsupervised learning has
 exploded in the last decade. One of the most popular generative models is the
 variational autoencoder (VAE) [@kingma2014]. VAEs assume the existence of a
 joint probability distribution between a high-dimensional data space and a
-lower-dimensional latent space. The VAE parametrizes this joint distribution
-with two neural networks--an encoder and a decoder--using a variational
-inference approach. This approach allows for the model to approximate the
-underlying low-dimensional structure that generated the observed data and
-generate new data samples by sampling from the learned latent space. Several
-variations of the original VAE have been proposed to extend its capabilities and
-tackle different problems. Here, we present `AutoEncoderToolkit.jl`, a `Julia`
-package for training VAEs and its extensions. The package is built on top of the
-`Flux.jl` deep learning library [@innes2018] and provides a simple and flexible
-interface for training different flavors of VAEs. Furthermore, the package
-provides a set of utilities for the geometric analysis of the learned latent
-space.
+lower-dimensional latent space. Using a variational inference approach, the VAE
+parametrizes this joint distribution with two neural networks--an encoder and a
+decoder. This approach allows the model to approximate the underlying
+low-dimensional structure that generated the observed data and generate new
+samples by sampling from the learned latent space. Several variations of the
+original VAE have been proposed to extend its capabilities and tackle different
+problems. Here, we present `AutoEncoderToolkit.jl`, a `Julia` package for
+training VAEs and its extensions. The package is built on top of the `Flux.jl`
+deep learning library [@innes2018] and provides a simple and flexible interface
+for training different flavors of VAEs. Furthermore, the package offers a set of
+utilities for the geometric analysis of the learned latent space.
 
 # Statement of need
 
-The collection and analysis of large high-dimensional datasets have become
-routine in several scientific fields. Therefore, the need to understand and
-uncover the underlying low-dimensional structure of these datasets is more
-pressing than ever. VAEs, have shown great promise in this regard, with
-applications ranging from single-cell transcriptomics [@lopez2018], to protein
-design [@lian2022], to the discovery of the governing equations of dynamical
-systems [@champion2019]. However, most of these tools have been exclusively
-developed for the `Python` ecosystem. `Julia` is a promising high-performance
-language for scientific computing, with a growing ecosystem for deep learning,
+Collecting and analyzing large high-dimensional datasets have become routine in
+several scientific fields. Therefore, the need to understand and uncover the
+underlying low-dimensional structure of these datasets has become more pressing
+than ever. VAEs have shown great promise in this regard, with applications
+ranging from single-cell transcriptomics [@lopez2018] to protein design
+[@lian2022] to the discovery of the governing equations of dynamical systems
+[@champion2019]. However, most of these tools have been exclusively developed
+for the `Python` ecosystem. `Julia` is a promising high-performance language for
+scientific computing, with a growing ecosystem for deep learning,
 state-of-the-art automatic differentiation, and a strong GPU-accelerated
 computing backend. Furthermore, the programming paradigm of `Julia`, based on
 multiple dispatch, allows for a more flexible, modular, and composable codebase.
 `AutoEncoderToolkit.jl` aims to provide a simple and flexible interface for
-training VAEs and its extensions in `Julia` taking full advantage of the
-programming paradigm of the language.
+training VAEs and their extensions in `Julia`, taking full advantage of the
+language programming paradigm.
 
 # Software Description
 
 ## Encoder \& Decoder definition
 
[…]
