### 2024-07-23 #6794 @mrazomej
@fabian-s, thanks for the information. Here is the Zenodo information:
```
DOI: 10.5281/zenodo.12802504
```
I also committed a new version of the manuscript in the `publication` branch with minor changes to the text, especially thanking the reviewers for their generous contribution to the project's quality.

I mentioned earlier in this issue that I couldn't find the corresponding DOIs for those two publications. Do you have any recommendations on what to do for that issue?
