### 2024-07-29 #6794 @fabian-s
sorry for the delay - it's exam grading season in Germany :crying_cat_face:
everything seems fine, non-existent DOIs are not a problem.
