### 2024-07-29 #6794 @fabian-s
@openjournals/joss-eics we checked that DOIs are not available where missing.
