### 2024-07-29 #6794 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
