### 2024-07-29 #6794 @danielskatz
@fabian-s - the <>s shouldn't be part of the tag, and they aren't in the source repo
