### 2024-07-29 #6794 @crvernon
:wave: @mrazomej - Well done! I just need you to address the following before I move to accept this submission for publication:

In the paper:
- [ ] Please remove the correspondence statement which has been set in your paper as an additional affiliation.  You can simply designate yourself as the corresponding author in your name attributes (see https://joss.readthedocs.io/en/latest/example_paper.html#example-paper).  This will not show your email address; however, your GitHub handle is shown in the final publication and you can add any personal information in your GitHub profile that you feel is necessary.
- [ ] Please add a URL and any other identifiable info to your Higgins et al. 2017 reference.  I know this is a conference paper that was in review but it would be good to have some other identifiable info with the record.
