### 2024-07-30 f10aa56
implement changes prior to publication
