+++ paper.bib
@@ -1,144 +1,145 @@
 @online{arvanitidis2021,
-  title = {Latent {{Space Oddity}}: On the {{Curvature}} of {{Deep Generative Models}}},
-  shorttitle = {Latent {{Space Oddity}}},
-  author = {Arvanitidis, Georgios and Hansen, Lars Kai and Hauberg, Søren},
-  date = {2021-12-13},
-  eprint = {1710.11379},
-  eprinttype = {arxiv},
+  title       = {Latent {{Space Oddity}}: On the {{Curvature}} of {{Deep Generative Models}}},
+  shorttitle  = {Latent {{Space Oddity}}},
+  author      = {Arvanitidis, Georgios and Hansen, Lars Kai and Hauberg, Søren},
+  date        = {2021-12-13},
+  eprint      = {1710.11379},
+  eprinttype  = {arXiv},
   eprintclass = {stat},
-  doi = {10.48550/arXiv.1710.11379},
-  pubstate = {preprint}
+  doi         = {10.48550/arXiv.1710.11379},
+  pubstate    = {prepublished}
 }
 
 @article{caterini2018,
-  title = {Hamiltonian {{Variational Auto-Encoder}}},
+  title  = {Hamiltonian {{Variational Auto-Encoder}}},
   author = {Caterini, Anthony L and Doucet, Arnaud and Sejdinovic, Dino},
-  date = {2018},
-  pages = {11},
-  doi = {10.48550/arXiv.1805.11328}
+  date   = {2018},
+  pages  = {11},
+  doi    = {10.48550/arXiv.1805.11328}
 }
 
 @online{chadebec2020,
-  title = {Geometry-{{Aware Hamiltonian Variational Auto-Encoder}}},
-  author = {Chadebec, Clément and Mantoux, Clément and Allassonnière, Stéphanie},
-  date = {2020-10-22},
-  eprint = {2010.11518},
-  eprinttype = {arxiv},
+  title       = {Geometry-{{Aware Hamiltonian Variational Auto-Encoder}}},
+  author      = {Chadebec, Clément and Mantoux, Clément and Allassonnière, Stéphanie},
+  date        = {2020-10-22},
+  eprint      = {2010.11518},
+  eprinttype  = {arXiv},
   eprintclass = {cs, math, stat},
-  doi = {10.48550/arXiv.2010.11518},
-  pubstate = {preprint}
+  doi         = {10.48550/arXiv.2010.11518},
+  pubstate    = {prepublished}
 }
 
 @online{chadebec2022,
-  title = {A {{Geometric Perspective}} on {{Variational Autoencoders}}},
-  author = {Chadebec, Clément and Allassonnière, Stéphanie},
-  date = {2022-11-03},
-  eprint = {2209.07370},
-  eprinttype = {arxiv},
+  title       = {A {{Geometric Perspective}} on {{Variational Autoencoders}}},
+  author      = {Chadebec, Clément and Allassonnière, Stéphanie},
+  date        = {2022-11-03},
+  eprint      = {2209.07370},
+  eprinttype  = {arXiv},
   eprintclass = {cs, stat},
-  doi = {10.48550/arXiv.2209.07370},
-  pubstate = {preprint}
+  doi         = {10.48550/arXiv.2209.07370},
+  pubstate    = {prepublished}
 }
 
 @article{champion2019,
-  title = {Data-Driven Discovery of Coordinates and Governing Equations},
-  author = {Champion, Kathleen and Lusch, Bethany and Kutz, J. Nathan and Brunton, Steven L.},
-  date = {2019-11-05},
+  title        = {Data-Driven Discovery of Coordinates and Governing Equations},
[…]
