### 2024-07-30 #6794 @mrazomej
@crvernon thank you very much!

I implemented both changes, so, in principle, the paper should be ready for publication. Please let me know if there is anything else I should change.

Thanks to the entire JOSS team. This has been, by far, the most pleasing peer-review process I have gone through. The way things are done for this journal gives me hope that there are ways to fix the broken review system.
